Fixes some warnings.

+++ json_utils.cpp
@@ -13,7 +13,7 @@ void to_json(json& j, menu_elem const& e)
 void from_json(json const& j, menu_elem& e)
 {
   e.data = j.at("data").get<std::string>();
-  e.depth = j.at("depth").get<unsigned long>();
+  e.depth = j.at("depth").get<unsigned>();
   e.depth = j.at("version").get<int>();
 }
